requirements.txt
asgiref==3.2.7
Django==2.2.10
djangorestframework==3.11.0
gunicorn==20.0.4
Pillow==7.1.2
psycopg2==2.8.5
pytz==2019.3
sqlparse==0.3.1
